+++ Changelog.md
@@ -12,6 +12,7 @@ Released on ?.
 
 - Fixed a crash when opening the Current Position page directly after app
   startup.
+- Fixed "Add new..." weather icon color in dark mode
 
 ## Version 1.13.0 "FJällräven Classic 2022" edition