+++ Changelog.md
@@ -6,6 +6,7 @@ Released on ?.
 
 ✨ New Features / Improvements
 
+- The "find result" pin can now be hidden again
 - using CesiumJS version 1.98.1
 
 🐛 Bugfixes