+++ Changelog.md
@@ -8,7 +8,7 @@ Released on ?.
 
 - The "find result" pin can now be hidden again
 - Added "altitude offset" to Flying Range dialog
-- using CesiumJS version 1.98.1
+- using CesiumJS version 1.99
 
 🐛 Bugfixes